--- nextstep-mvc/src/main/java/nextstep/mvc/tobe/NotFoundControllerException.java
@@ -1,9 +1,0 @@
-package nextstep.mvc.tobe;
-
-import javassist.NotFoundException;
-
-public class NotFoundControllerException extends NotFoundException {
-    public NotFoundControllerException(String msg) {
-        super(msg);
-    }
-}
