+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/HandlerExecutionHandlerAdapter.java
@@ -1,5 +1,7 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.adapter;
 
+import nextstep.mvc.HandlerAdapter;
+import nextstep.mvc.tobe.handler.HandlerExecution;
 import nextstep.mvc.tobe.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
