+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/NotFoundAdapterException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.adapter;
 
 import javassist.NotFoundException;
