+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/handler/NotFoundHandlerException.java
@@ -1,0 +1,9 @@
+package nextstep.mvc.tobe.handler;
+
+import javassist.NotFoundException;
+
+public class NotFoundHandlerException extends NotFoundException {
+    public NotFoundHandlerException(String msg) {
+        super(msg);
+    }
+}
