+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/JspView.java
@@ -1,5 +1,7 @@
 package nextstep.mvc.tobe.view;
 
+import nextstep.mvc.View;
+
 import javax.servlet.RequestDispatcher;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
